feat(tests): enhance player and world functionality tests
- Added tests for player respawn behavior, ensuring all relevant state resets occur.
- Implemented checks for enemy removal during player kick actions to prevent skipping.
- Introduced new tests for region graph extraction, validating body-blocking materials.
- Enhanced runtime panel tests to include accessibility attributes.
- Added world generation tests to ensure deterministic behavior and validate secret connector trials.
- Created input manager tests to verify gameplay key codes.
- Implemented level store tests to validate RLE save application and error handling.
- Added RLE codec tests to ensure correct encoding and decoding behavior.
- Introduced modal focus trap utility for managing focus within modals.
- Added scripts for runtime UI verification and improved test execution handling.

+++ src/render/ComposeShader.ts
@@ -334,7 +334,7 @@ void main() {
 function packCellValue(
   types: Uint8Array,
   colors: Uint32Array,
-  charge: Uint8Array,
+  charge: Uint16Array,
   ci: number,
 ): number {
   const c = colors[ci];
@@ -813,7 +813,7 @@ export class GpuCompose {
   private packWindowRows(
     types: Uint8Array,
     colors: Uint32Array,
-    charge: Uint8Array,
+    charge: Uint16Array,
     camX: number,
     camY: number,
     startRow: number,
